@@ -1,6 +1,6 @@
 package com.steelkiwi.cropiwa;
 
-import android.support.annotation.IntRange;
+import androidx.annotation.IntRange;
 
 /**
  * Created by yarolegovich https://github.com/yarolegovich
