@@ -2,7 +2,8 @@ package com.steelkiwi.cropiwa.config;
 
 import android.graphics.Bitmap;
 import android.net.Uri;
-import android.support.annotation.IntRange;
+
+import androidx.annotation.IntRange;
 
 import com.steelkiwi.cropiwa.image.CropIwaBitmapManager;
 /**
