@@ -6,8 +6,9 @@ import android.graphics.BitmapFactory;
 import android.graphics.Matrix;
 import android.media.ExifInterface;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.steelkiwi.cropiwa.config.CropIwaSaveConfig;
 import com.steelkiwi.cropiwa.shape.CropIwaShapeMask;
@@ -26,7 +27,7 @@ import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.steelkiwi.cropiwa.util.CropIwaUtils.*;
+import static com.steelkiwi.cropiwa.util.CropIwaUtils.closeSilently;
 
 /**
  * @author Yaroslav Polyakov https://github.com/polyak01
