@@ -2,9 +2,10 @@ package com.steelkiwi.cropiwa.config;
 
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.support.annotation.FloatRange;
-import android.support.annotation.NonNull;
 import android.util.AttributeSet;
+
+import androidx.annotation.FloatRange;
+import androidx.annotation.NonNull;
 
 import com.steelkiwi.cropiwa.AspectRatio;
 import com.steelkiwi.cropiwa.R;
